+++ extension/src/NABfunctions.ts
@@ -8,7 +8,6 @@ import * as Documentation from "./Documentation";
 import * as DebugTests from "./DebugTests";
 import * as ALParser from "./ALObject/ALParser";
 import * as path from "path";
-import * as Common from "./Common";
 import * as PowerShellFunctions from "./PowerShellFunctions";
 import * as DocumentFunctions from "./DocumentFunctions";
 import * as FileFunctions from "./FileFunctions";
@@ -17,7 +16,7 @@ import { xliffCache } from "./Xliff/XLIFFCache";
 import * as Telemetry from "./Telemetry";
 import * as PermissionSetFunctions from "./PermissionSet/PermissionSetFunctions";
 import { IOpenXliffIdParam } from "./Types";
-import { TargetState, TransUnit, Xliff } from "./Xliff/XLIFFDocument";
+import { TargetState, Xliff } from "./Xliff/XLIFFDocument";
 import { baseAppTranslationFiles } from "./externalresources/BaseAppTranslationFiles";
 import { XliffEditorPanel } from "./XliffEditor/XliffEditorPanel";
 import * as fs from "fs";
@@ -38,7 +37,7 @@ import { TextDocumentMatch } from "./Types";
 import { logger } from "./Logging/LogHelper";
 import { PermissionSetNameEditorPanel } from "./PermissionSet/PermissionSetNamePanel";
 import { TemplateEditorPanel } from "./Template/TemplatePanel";
-import { OutputLogger } from "./Logging/OutputLogger";
+import { showErrorAndLog } from "./ErrorMessage";
 
 export async function refreshXlfFilesFromGXlf(
   suppressMessage = false
@@ -587,14 +586,6 @@ export async function generateExternalDocumentation(): Promise<void> {
   }
 
   logger.log("Done: GenerateExternalDocumentation");
-}
-
-function showErrorAndLog(action: string, error: Error): void {
-  const errMsg = `${action} failed with error: ${error.message}`;
-  vscode.window.showErrorMessage(errMsg);
-  logger.log(`Error: ${error.message}`);
-  logger.log(`Stack trace: ${error.stack}`);
-  Telemetry.trackException(error);
 }
 
 export async function matchTranslations(): Promise<void> {
@@ -1385,99 +1376,6 @@ function appendActiveDocument(filesToSearch: string[]): string[] {
   }
   return filesToSearch;
 }
-export function troubleshootParseCurrentFile(): void {
-  logger.log("Running: troubleshootParseCurrentFile");
-  Telemetry.trackEvent("troubleshootParseCurrentFile");
-  try {
-    const currDocument = vscode.window.activeTextEditor?.document;
-    if (!currDocument) {
-      throw new Error("This command must be run with an open editor.");
-    }
-    if (!currDocument.getText()) {
-      throw new Error("This command must be run with an open editor.");
-    }
-    const alObj = ALParser.getALObjectFromText(currDocument.getText(), true);
-    if (!alObj) {
-      throw new Error("No object descriptor was found in the open editor.");
-    }
-    logger.log("---------------------------------------------");
-    logger.log(`Object Type: ${alObj.objectType}`);
-    logger.log(`Object Id: ${alObj.objectId}`);
-    logger.log(`Object Name: ${alObj.objectName}`);
-    if (alObj.extendedObjectId) {
-      logger.log(`Extends Object ID: ${alObj.extendedObjectId}`);
-    }
-    if (alObj.extendedObjectName) {
-      logger.log(`Extends Object Name: ${alObj.extendedObjectName}`);
-    }
-    if (alObj.extendedTableId) {
-      logger.log(`TableId of Extended page: ${alObj.extendedTableId}`);
-    }
-    logger.log();
-    logger.log("AL Object properties:");
-    alObj.properties.forEach((p) => logger.log(`${p.name} = ${p.value}`));
-    logger.log();
-    logger.log("Multi Language Controls:");
-    alObj
-      .getAllMultiLanguageObjects()
-      .forEach((m) => logger.log(`${m.name} = ${m.text}`));
-    logger.log();
-    logger.log("Controls:");
-    alObj.getAllControls().forEach((c) => logger.log(`${c.type}: ${c.name}`));
-    logger.log();
-
-    alObj.prepareForJsonOutput();
-    vscode.workspace
-      .openTextDocument({
-        language: "json",
-        content: Common.orderedJsonStringify(alObj, 4),
-      })
-      .then((doc) => vscode.window.showTextDocument(doc));
-    vscode.window.showInformationMessage(
-      `The .al file was successfully parsed. Open the Output channel '${OutputLogger.channelName}' for details. Review the opened json file for the parsed object structure.`
-    );
-  } catch (error) {
-    showErrorAndLog(
-      "Parsing of current AL Object failed with error:",
-      error as Error
-    );
-  }
-  logger.show();
-}
-export async function troubleshootParseAllFiles(): Promise<void> {
-  logger.log("Running: troubleshootParseAllFiles");
-  Telemetry.trackEvent("troubleshootParseAllFiles");
-  try {
-    const objects = await WorkspaceFunctions.getAlObjectsFromCurrentWorkspace(
-      SettingsLoader.getSettings(),
-      SettingsLoader.getAppManifest(),
-      true
-    );
-    for (const alObj of objects) {
-      alObj.prepareForJsonOutput();
-    }
-    vscode.workspace
-      .openTextDocument({
-        language: "json",
-        content: Common.orderedJsonStringify(objects, 4),
-      })
-      .then((doc) => vscode.window.showTextDocument(doc));
-    logger.log();
-    logger.log("Objects:");
-    objects.forEach((obj) =>
-      logger.log(`${obj.objectType} ${obj.objectId} ${obj.objectName}`)
-    );
-    vscode.window.showInformationMessage(
-      `All .al file was successfully parsed. Review the opened json file for the parsed object structure. Any missing object could not be identified as an AL object, please report as an issue on GitHub (https://github.com/jwikman/nab-al-tools/issues)`
-    );
-  } catch (error) {
-    showErrorAndLog(
-      "Parsing of all AL Objects failed with error:",
-      error as Error
-    );
-  }
-  logger.show();
-}
 export async function createProjectFromTemplate(
   extensionUri: vscode.Uri
 ): Promise<void> {
@@ -1501,7 +1399,7 @@ export async function createProjectFromTemplate(
       async (workspaceFilePath) => {
         if (workspaceFilePath !== "") {
           if (fs.existsSync(workspaceFilePath)) {
-            logger.log("Open workspacefile: ", workspaceFilePath);
+            logger.log("Open workspace file: ", workspaceFilePath);
             const uri = vscode.Uri.file(workspaceFilePath);
             await vscode.commands.executeCommand("vscode.openFolder", uri);
           }
@@ -1534,76 +1432,5 @@ export async function renumberALObjects(): Promise<void> {
     logger.log("Done: renumberALObjects");
   } catch (error) {
     showErrorAndLog("Renumber AL objects", error as Error);
-  }
-}
-export async function troubleshootFindTransUnitsWithoutSource(): Promise<void> {
-  logger.log("Running: troubleshootFindTransUnitsWithoutSource");
-  Telemetry.trackEvent("troubleshootFindTransUnitsWithoutSource");
-
-  const settings = SettingsLoader.getSettings();
-  const appManifest = SettingsLoader.getAppManifest();
-  const gXlfFilePath = WorkspaceFunctions.getGXlfFilePath(
-    settings,
-    appManifest
-  );
-  const alObjects = await WorkspaceFunctions.getAlObjectsFromCurrentWorkspace(
-    settings,
-    appManifest,
-    true
-  );
-  const gXlf = Xliff.fromFileSync(gXlfFilePath);
-  const failingTransUnits: TransUnit[] = [];
-  gXlf.transunit.forEach((tu) => {
-    try {
-      const obj = DocumentFunctions.getObjectFromTokens(
-        alObjects,
-        tu.getXliffIdTokenArray()
-      );
-      const mlObjects = obj.getAllMultiLanguageObjects({
-        onlyForTranslation: true,
-      });
-      const mlObject = mlObjects.find(
-        (x) => x.xliffId().toLowerCase() === tu.id.toLowerCase()
-      );
-      if (!mlObject) {
-        throw new Error("ML Object not found");
-      }
-      // console.log(`${tu.id} => ${mlObject.text}`);
-    } catch (error) {
-      console.log(`Missing source: "${tu.id}"`);
-      failingTransUnits.push(tu);
-    }
-  });
-  if (failingTransUnits.length === 0) {
-    vscode.window.showInformationMessage(
-      "All is OK, could find the source of all TransUnits."
-    );
-  } else {
-    logger.error("The following TransUnits could not identify it's code:");
-    failingTransUnits.forEach((tu) =>
-      logger.log(`${tu.id} - ${tu.xliffGeneratorNoteContent()}`)
-    );
-    const result = LanguageFunctions.findTransUnitId(
-      failingTransUnits[0].id,
-      fs.readFileSync(gXlfFilePath, "utf8"),
-      gXlfFilePath
-    );
-    if (result) {
-      DocumentFunctions.openTextFileWithSelection(
-        result.filePath,
-        result.position,
-        result.length
-      );
-      vscode.window.showErrorMessage(
-        `There was ${failingTransUnits.length} TransUnits that could find it's source. Investigate the NAB AL Tools log to identify them. The first of them is now opened.`,
-        { modal: true }
-      );
-    } else {
-      vscode.window.showErrorMessage(
-        `There was ${failingTransUnits.length} TransUnits that could find it's source. Investigate the NAB AL Tools log to identify them.`,
-        { modal: true }
-      );
-    }
-    logger.show();
   }
 }
